refactor: moved bugfix into adapter pattern
