+++ lib/Client/index.ts
@@ -2,6 +2,7 @@ import { State } from '@voiceflow/runtime';
 import { AxiosInstance } from 'axios';
 
 import { RequestContext, ResponseContext } from '@/lib/types';
+import { adaptResponseContext } from "./adapters";
 
 class Client {
   private axios: AxiosInstance;
@@ -15,7 +16,9 @@ class Client {
   }
 
   async interact(body: RequestContext, versionID: string): Promise<ResponseContext> {
-    return this.axios.post(`/interact/${versionID}`, body).then((response) => response.data);
+    return this.axios.post(`/interact/${versionID}`, body)
+      .then((response) => response.data)
+      .then((context) => adaptResponseContext(context));
   }
 }
